Update DataCleaningScript.R
Now that employer size is a factor variable, the old approach does not work.  After looking up the error I realized that the issue was that the levels I wanted had to be defined.  Initially, I thought I would have to reorder them and then code it again, but when I reordered them it messed it up.  Then I started from scratch and just replaced the odd factor levels with the correct ones and it worked!